refactor(cli): remove useless file
affects: @fect-ui/cli, @fect-ui/vue-icons
