+++ packages/cli/src/commands/collect.ts
@@ -1,4 +1,4 @@
-import { genPackagesEntry } from '../compiler/collect'
+import { genPackagesEntry } from '../node/collect'
 
 export const collect = async () => {
   await genPackagesEntry()
