--- packages/cli/src/commands/compile.ts
@@ -1,5 +1,0 @@
-import { bundler } from '../compiler/bundler'
-
-export const compile = () => {
-  bundler()
-}
