+++ packages/cli/src/commands/index.ts
@@ -1,4 +1,3 @@
-export * from './compile'
 export * from './collect'
 export * from './test'
 export * from './volar'
