+++ packages/cli/src/index.ts
@@ -1,1 +1,2 @@
 export { defineNonConfig } from './node/config'
+export { compile } from './node/compile'
