+++ packages/cli/src/shared/format.ts
